[MOVE] some method to static itemstack factory
