+++ src/main/java/com/github/unchama/seichiassist/data/ItemData.java
@@ -1,6 +1,6 @@
 package com.github.unchama.seichiassist.data;
 
-import com.github.unchama.seichiassist.util.Util;
+import com.github.unchama.seichiassist.util.StaticGachaPrizeFactory;
 import com.github.unchama.util.collection.ImmutableListFactory;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -50,8 +50,8 @@ public class ItemData {
 		ItemMeta meta;
 		gachaimo = new ItemStack(Material.GOLDEN_APPLE, amount);
 		meta = Bukkit.getItemFactory().getItemMeta(Material.GOLDEN_APPLE);
-		meta.setDisplayName(Util.getGachaRingoName());
-		List<String> lore = Util.getGachaRingoLore();
+		meta.setDisplayName(StaticGachaPrizeFactory.getGachaRingoName());
+		List<String> lore = StaticGachaPrizeFactory.getGachaRingoLore();
 		meta.setLore(lore);
 		gachaimo.setItemMeta(meta);
 		return gachaimo;
