+++ src/main/java/com/github/unchama/seichiassist/effect/arrow/ArrowExplosionTask.java
@@ -1,20 +1,18 @@
 package com.github.unchama.seichiassist.effect.arrow;
 
-import java.util.HashMap;
-import java.util.UUID;
-
+import com.github.unchama.seichiassist.SeichiAssist;
+import com.github.unchama.seichiassist.data.PlayerData;
 import org.bukkit.Location;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.SmallFireball;
 import org.bukkit.metadata.FixedMetadataValue;
-import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.util.Vector;
 
-import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.data.PlayerData;
+import java.util.HashMap;
+import java.util.UUID;
 
-public class ArrowExplosionTask extends BukkitRunnable{
+public class ArrowExplosionTask extends SkeletonEffectTask<SmallFireball> {
 	SeichiAssist plugin = SeichiAssist.instance;
 	HashMap<UUID,PlayerData> playermap = SeichiAssist.playermap;
 	Player player;
@@ -22,7 +20,6 @@ public class ArrowExplosionTask extends BukkitRunnable{
 	UUID uuid;
 	PlayerData playerdata;
 	long tick;
-	SmallFireball proj;
 
 	public ArrowExplosionTask(Player player) {
 		this.tick = 0;
@@ -35,7 +32,7 @@ public class ArrowExplosionTask extends BukkitRunnable{
 		this.playerdata = playermap.get(uuid);
 
 		//発射する音を再生する.
-		player.playSound(ploc, Sound.ENTITY_GHAST_SHOOT, 1, (float)1.3);
+		player.playSound(ploc, Sound.ENTITY_GHAST_SHOOT, 1.0f, 1.3f);
 
 		//スキルを実行する処理
 		Location loc = player.getLocation().clone();
@@ -65,7 +62,7 @@ public class ArrowExplosionTask extends BukkitRunnable{
 		if(tick > 100){
 			proj.remove();
 			SeichiAssist.entitylist.remove(proj);
-			this.cancel();
+			cancel();
 		}
 	}
