+++ src/main/java/com/github/unchama/seichiassist/effect/breaking/MagicTask.java
@@ -1,14 +1,11 @@
 package com.github.unchama.seichiassist.effect.breaking;
 
-import java.util.List;
-import java.util.Random;
-
-import org.bukkit.DyeColor;
-import org.bukkit.EntityEffect;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.Particle;
-import org.bukkit.Sound;
+import com.github.unchama.seichiassist.SeichiAssist;
+import com.github.unchama.seichiassist.data.Coordinate;
+import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.task.AsyncEntityRemover;
+import com.github.unchama.seichiassist.util.BreakUtil;
+import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockState;
 import org.bukkit.entity.Chicken;
@@ -16,16 +13,11 @@ import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.material.Wool;
-import org.bukkit.scheduler.BukkitRunnable;
 
-import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.data.Coordinate;
-import com.github.unchama.seichiassist.data.PlayerData;
-import com.github.unchama.seichiassist.task.AsyncEntityRemover;
-import com.github.unchama.seichiassist.util.BreakUtil;
+import java.util.List;
+import java.util.Random;
 
-public class MagicTask extends BukkitRunnable {
-	SeichiAssist plugin = SeichiAssist.instance;
+public class MagicTask extends SkeletonBreakTask {
 	// プレイヤー情報
 	Player player;
 	// プレイヤーデータ
@@ -33,7 +25,6 @@ public class MagicTask extends BukkitRunnable {
 	// プレイヤーの位置情報
 	Location ploc;
 	// ブロックの位置情報
-	Location bloc;
 	// 破壊するブロックの中心位置
 	Location centerbreakloc;
 	// 使用するツール
@@ -71,49 +62,42 @@ public class MagicTask extends BukkitRunnable {
 	}
 
 	@Override
-	public void run() {
-		round++;
-		switch(round){
-		case 1:
-			//1回目のrun
+	public void callFirst() {
+		DyeColor[] colors = { DyeColor.RED, DyeColor.BLUE, DyeColor.YELLOW, DyeColor.GREEN };
+		int rd = new Random().nextInt(colors.length);
 
-			DyeColor[] colors = { DyeColor.RED, DyeColor.BLUE, DyeColor.YELLOW, DyeColor.GREEN };
-			int rd = new Random().nextInt(colors.length);
+		for (Block b : breaklist) {
+			BreakUtil.breakBlock(player, b, droploc, tool, false);
+			b.setType(Material.WOOL);
+			state = b.getState();
+			red = (Wool) state.getData();
+			red.setColor(colors[rd]);
+			state.update();
+		}
+	}
 
-			for (Block b : breaklist) {
-				BreakUtil.breakBlock(player, b, droploc, tool, false);
-				b.setType(Material.WOOL);
-				state = b.getState();
-				red = (Wool) state.getData();
-				red.setColor(colors[rd]);
-				state.update();
-			}
-			break;
-
-		case 2:
-			//2回目のrun
-			if (SeichiAssist.entitylist.isEmpty()) {
-				Chicken e = (Chicken) player.getWorld().spawnEntity(centerbreakloc, EntityType.CHICKEN);
-				SeichiAssist.entitylist.add(e);
-				e.playEffect(EntityEffect.WITCH_MAGIC);
-				e.setInvulnerable(true);
-				new AsyncEntityRemover(e).runTaskLater(plugin, 100);
-				player.getWorld().playSound(effectloc, Sound.ENTITY_WITCH_AMBIENT, 1, 1.5F);
-			}
-
-			for (Block b : breaklist) {
-				b.setType(Material.AIR);
-				b.getWorld().spawnParticle(Particle.NOTE, b.getLocation().add(0.5, 0.5, 0.5), 1);
-				SeichiAssist.allblocklist.remove(b);
-			}
-			cancel();
-			break;
-
-		default:
-			//念のためキャンセル
-			cancel();
-			break;
+	@Override
+	public void callSecond() {
+		if (SeichiAssist.entitylist.isEmpty()) {
+			Chicken e = (Chicken) player.getWorld().spawnEntity(centerbreakloc, EntityType.CHICKEN);
+			SeichiAssist.entitylist.add(e);
+			e.playEffect(EntityEffect.WITCH_MAGIC);
+			e.setInvulnerable(true);
+			new AsyncEntityRemover(e).runTaskLater(SeichiAssist.instance, 100);
+			player.getWorld().playSound(effectloc, Sound.ENTITY_WITCH_AMBIENT, 1, 1.5F);
 		}
 
+		for (Block b : breaklist) {
+			b.setType(Material.AIR);
+			b.getWorld().spawnParticle(Particle.NOTE, b.getLocation().add(0.5, 0.5, 0.5), 1);
+			SeichiAssist.allblocklist.remove(b);
+		}
+		cancel();
+	}
+
+	@Override
+	public void callAfterThird() {
+		//念のためキャンセル
+		cancel();
 	}
 }
